+++ CleanArchitecture/Tests/MockContainer.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+class MockContainer {
+    var view:MockView?
+}