+++ CleanArchitecture/Tests/MockView.swift
@@ -1,11 +1,11 @@
 import Foundation
 import CleanArchitecture
 
-class MockView:View<MockPresenterProtocol, MockUIView> {
+class MockView:View<MockPresenter> {
     var onPropertyUpdated:(() -> Void)?
     
     func startObserving() {
-        self.presenter.viewModel.observe { (property:MockViewModelProtocol) in
+        self.presenter.viewModels.observe { (property:MockViewModel) in
             self.onPropertyUpdated?()
         }
     }