+++ CleanArchitecture/Tests/TestView.swift
@@ -6,19 +6,13 @@ class TestView:XCTestCase {
         let view:MockView = MockView()
         XCTAssertNotNil(view.presenter, "Not injected")
         XCTAssertNotNil(view.presenter.interactor, "Not injected")
-        XCTAssertNotNil(view.presenter.interactor.presenter, "Not injected")
-        XCTAssertNotNil(view.presenter.viewModel, "Not injected")
-    }
-    
-    func testViewContent() {
-        let view:MockView = MockView()
-        let content:MockUIView? = view.view as? MockUIView
-        XCTAssertNotNil(content, "Invalid content type")
+        XCTAssertNotNil(view.presenter.interactor.delegate, "Not injected")
+        XCTAssertNotNil(view.presenter.viewModels, "Not injected")
     }
     
     func testCallsPresenterOnViewDidLoad() {
-        let presenter:MockPresenterProtocol = MockPresenterProtocol()
-        presenter.interactor = MockInteractorProtocol()
+        let presenter:MockPresenter = MockPresenter()
+        presenter.interactor = MockInteractor()
         let view:MockView = MockView(presenter:presenter)
         var called:Bool = false
         presenter.onDidLoad = { called = true }
@@ -27,8 +21,8 @@ class TestView:XCTestCase {
     }
     
     func testCallsInteractorOnViewDidLoad() {
-        let presenter:MockPresenterProtocol = MockPresenterProtocol()
-        presenter.interactor = MockInteractorProtocol()
+        let presenter:MockPresenter = MockPresenter()
+        presenter.interactor = MockInteractor()
         let view:MockView = MockView(presenter:presenter)
         var called:Bool = false
         presenter.interactor.onDidLoad = { called = true }
@@ -37,7 +31,7 @@ class TestView:XCTestCase {
     }
     
     func testCallsPresenterOnViewWillAppear() {
-        let presenter:MockPresenterProtocol = MockPresenterProtocol()
+        let presenter:MockPresenter = MockPresenter()
         let view:MockView = MockView(presenter:presenter)
         var called:Bool = false
         presenter.onWillAppear = { called = true }
@@ -46,22 +40,11 @@ class TestView:XCTestCase {
     }
     
     func testCallsPresenterOnViewDidAppear() {
-        let presenter:MockPresenterProtocol = MockPresenterProtocol()
+        let presenter:MockPresenter = MockPresenter()
         let view:MockView = MockView(presenter:presenter)
         var called:Bool = false
         presenter.onDidAppear = { called = true }
         view.viewDidAppear(false)
         XCTAssertTrue(called, "Not called")
-    }
-    
-    func testInjectContentWithPresenter() {
-        let presenter:MockPresenterProtocol = MockPresenterProtocol()
-        let view:MockView = MockView(presenter:presenter)
-        XCTAssertNotNil(view.content, "Not injected")
-    }
-    
-    func testInjectContentWithoutPresenter() {
-        let view:MockView = MockView()
-        XCTAssertNotNil(view.content, "Not injected")
     }
 }